alx_travel_app/requirements.txt
aiosqlite==0.21.0
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-seed==0.3.1
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.10
Faker==37.8.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
mysql-connector-python==9.4.0
mysqlclient==2.2.7
packaging==25.0
parameterized==0.9.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
pycodestyle==2.14.0
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rabbitmq==0.2.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
